+++ CHANGELOG.md
@@ -8,16 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 ### Added
-- context: add `Context::set_restart_policy` to override a group's default restart policy.
+- context: add `Context::set_restart_policy` to override a group's default restart policy ([#48]).
 
 ### Changed
-- **BREAKING** group: make the module private, reexport `TerminationPolicy` and `RestartPolicy`.
+- **BREAKING** group: make the module private, reexport `TerminationPolicy` and `RestartPolicy` ([#48]).
 - **BREAKING** core: `start` and `try_start` functions are moved to `elfo::init` module ([#95]).
 
 ### Added
 - init,configurer: allow to check configs without starting the system ([#95]).
 
 [#95]: https://github.com/elfo-rs/elfo/pull/95
+[#48]: https://github.com/elfo-rs/elfo/issues/48
 
 ## [0.2.0-alpha.2] - 2023-06-14
 ### Added